Ananya_Contributions/object_detection/rcnn_flask_api/Pothole_Detection/requirements.txt
numpy
torch
torchvision
pybase64
bbox-utils
Flask
Pillow
gunicorn
flask-cors
flask-limiter
flask-restful

